+++ src/components/BookForm.js
@@ -1,0 +1,10 @@
+import React from 'react';
+
+function BookForm() {
+  return <div>
+    <input type="text" name="addBooks" placeholder="Book Title" />
+    <button type="button">Add Books</button>
+  </div>;
+}
+
+export default BookForm;
