+++ src/components/Books.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import BookForm from './BookForm';
+import Booklist from './Booklist';
+
+function Books() {
+  return (
+    <div>
+      <Booklist />
+      <BookForm />
+    </div>
+  );
+}
+
+export default Books;
