+++ src/redux/books/books.js
@@ -18,7 +18,7 @@ const reducer = (state = initialState, action) => {
     case ADD_BOOK:
       return [...state, action.payload];
     case REMOVE_BOOK:
-      return state.filter(book => book.id !== id);
+      return state.filter((book) => book.id !== action.payload.id);
     default:
       return state;
   }
